Rename: entity package to domain
