+++ src/main/java/org/prgrms/kdt/CommandLineApplication.java
@@ -1,7 +1,7 @@
 package org.prgrms.kdt;
 
 import org.prgrms.kdt.core.*;
-import org.prgrms.kdt.entity.VoucherType;
+import org.prgrms.kdt.domain.VoucherType;
 import org.prgrms.kdt.service.*;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
