+++ src/main/java/org/prgrms/kdt/OrderTester.java
@@ -1,6 +1,6 @@
 package org.prgrms.kdt;
 
-import org.prgrms.kdt.entity.OrderItem;
+import org.prgrms.kdt.domain.OrderItem;
 import org.prgrms.kdt.service.OrderService;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.util.Assert;
