+++ src/main/java/org/prgrms/kdt/domain/VoucherType.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.entity;
+package org.prgrms.kdt.domain;
 
 public enum VoucherType {
     FIXED_AMOUNT,
