+++ src/main/java/org/prgrms/kdt/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package org.prgrms.kdt.repository;
 
-import org.prgrms.kdt.entity.Order;
+import org.prgrms.kdt.domain.Order;
 
 public interface OrderRepository {
     void insert(Order order);
